Change SignupPage to SignupForm
