+++ src/Routes.js
@@ -6,7 +6,6 @@ import ShowPage from './components/ShowPage';
 import WelcomePage from './components/WelcomePage';
 import CreatePage from './components/CreatePage';
 import LoginPage from './components/LoginPage';
-import SignupPage from './components/SignupPage';
 import DashboardPage from './components/DashboardPage';
 
 const Routes = () => (
@@ -18,7 +17,6 @@ const Routes = () => (
       <Route path="/show/:showId" component={ShowPage} />
       <Route path="/new" component={CreatePage} />
       <Route path="/login" component={LoginPage} />
-      <Route path="/signup" component={SignupPage} />
       <Route path="/dashboard" component={DashboardPage} />
     </Switch>
   </>
